refactor: move entites to aws folder (#736)
