+++ app/apis/catalog/hca-atlas-tracker/common/entities.ts
@@ -658,44 +658,3 @@ export enum ROLE {
   STAKEHOLDER = "STAKEHOLDER",
   UNREGISTERED = "UNREGISTERED",
 }
-
-// S3 and SNS Event Interfaces
-export interface S3Object {
-  eTag: string;
-  key: string;
-  size: number;
-  userMetadata?: {
-    "source-sha256": string;
-  };
-  versionId: string;
-}
-
-export interface S3EventRecord {
-  eventName: string;
-  eventSource: string;
-  eventTime: string;
-  eventVersion: string;
-  s3: {
-    bucket: {
-      name: string;
-    };
-    object: S3Object;
-    s3SchemaVersion: string;
-  };
-}
-
-export interface S3Event {
-  Records: S3EventRecord[];
-}
-
-export interface SNSMessage {
-  Message: string;
-  MessageId: string;
-  Signature: string;
-  SignatureVersion: string;
-  SigningCertURL: string;
-  Subject: string;
-  Timestamp: string;
-  TopicArn: string;
-  Type: string;
-}
